Move parser to separate file and fix lint errors
