+++ lib/emitter.js
@@ -1,14 +1,16 @@
-exports.document = function(doc) {
-  return doc.map(emit).join('');
-}
+'use strict';
 
-exports.fragment = function(doc) {
+exports.document = function (doc) {
+  return doc.map(emit).join('');
+};
+
+exports.fragment = function (doc) {
   if (doc.length === 0) {
     return '';
   }
 
   return doc[0].contents.map(emit).join('');
-}
+};
 
 function emit(node) {
   var str = '';
