+++ test/run-cases.js
@@ -9,20 +9,20 @@ Bluebird.try(function () {
     casesDirectory: path.resolve(__dirname, 'list-cases'),
     inputExtension: 'ecmarkdown',
     outputExtension: 'html'
-  })
+  });
 })
 .then(function () {
   return baselineTester(ecmarkdown.document, {
     casesDirectory: path.resolve(__dirname, 'paragraph-cases'),
     inputExtension: 'ecmarkdown',
     outputExtension: 'html'
-  })
+  });
 })
 .then(function () {
   return baselineTester(ecmarkdown.fragment, {
     casesDirectory: path.resolve(__dirname, 'fragment-cases'),
     inputExtension: 'ecmarkdown',
     outputExtension: 'html'
-  })
+  });
 })
 .done();
